replace identity api code from mvc project to identity api project
