+++ Frontend/SciMaterials.UI.MVC/appsettings.json
@@ -19,20 +19,5 @@
     "Encoding": "UTF-32",
     "ConcurrentDbTimeout": 100,
     "ConcurrentDbTryCount": 10
-  },
-  "IdentityDatabase": {
-    "ConnectionStrings": {
-      "SqlServer.Debug": "Server=localhost;Database=SciMaterials.Identity;User Id=sa;Password=veryStrongPassword123;"
-    },
-    "Provider": "SqlServer.Debug",
-    "RemoveAtStart": "true",
-    "UseDataSeeder": "true"
-  },
-  "IdentitySettings": {
-    "AdminSettings": {
-      "login": "sa@mail.ru",
-      "password": "test12345"
-    },
-    "SecretTokenKey": "vRguqNCvKJHkhs578443sa46@#sakjjRxDGMFguqNCMz8w2DGMQ=="
   }
 }
