+++ Identity/SciMaterials.Identity.Api/IdentityRegister.cs
@@ -3,17 +3,14 @@ using Microsoft.AspNetCore.Authentication.JwtBearer;
 using Microsoft.AspNetCore.Identity;
 using Microsoft.IdentityModel.Tokens;
 using Microsoft.OpenApi.Models;
-using SciMaterials.Contracts.Identity.API;
 using SciMaterials.DAL.AUTH.Context;
 using SciMaterials.DAL.AUTH.Contracts;
 using SciMaterials.DAL.AUTH.InitializationDb;
 using SciMaterials.DAL.Contracts.Configuration;
 using SciMaterials.Identity.DAL.SqlServer.Migrations;
-using SciMaterials.Services.Identity.API;
-using SciMaterials.UI.MVC.Identity.Services;
 using Swashbuckle.AspNetCore.SwaggerGen;
 
-namespace SciMaterials.UI.MVC;
+namespace SciMaterials.Identity.Api;
 
 public static class IdentityRegister
 {
@@ -116,21 +113,7 @@ public static class IdentityRegister
 
         return Services.AddSingleton<IAuthUtils, AuthUtils>();
     }
-    
-    /// <summary>Метод расширения для Identity клиента</summary>
-    /// <param name="services">Сервисы</param>
-    /// <param name="serverUrl">Url сервера</param>
-    /// <returns></returns>
-    public static IServiceCollection AddIdentityClients(this IServiceCollection services, string serverUrl)
-    {
-        services.AddHttpClient<IIdentityApi, IdentityClient>("IdentityApi", c =>
-        {
-            c.BaseAddress = new Uri(serverUrl);
-        });
-        
-        return services;
-    }
-    
+
     /// <summary>Метод расширения для инициализации базы данных Identity</summary>
     /// <param name="app">Интерфес IApplicationBuilder</param>
     public static async Task InitializeIdentityDatabaseAsync(this IApplicationBuilder app)
