+++ Identity/SciMaterials.Identity.Api/SciMaterials.Identity.Api.csproj
@@ -8,6 +8,15 @@
 
   <ItemGroup>
     <PackageReference Include="Swashbuckle.AspNetCore" Version="6.2.3" />
+    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="6.25.0" />
+    <PackageReference Include="Microsoft.AspNetCore.Authentication.JwtBearer" Version="6.0.10" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\Libraries\SciMaterials.Contracts.Identity.API\SciMaterials.Contracts.Identity.API.csproj" />
+    <ProjectReference Include="..\Libraries\SciMaterials.DAL.AUTH.Contracts\SciMaterials.DAL.AUTH.Contracts.csproj" />
+    <ProjectReference Include="..\Libraries\SciMaterials.Identity.DAL.SqlServer.Migrations\SciMaterials.Identity.DAL.SqlServer.Migrations.csproj" />
+    <ProjectReference Include="..\Libraries\SciMaterials.Services.Identity.API\SciMaterials.Services.Identity.API.csproj" />
   </ItemGroup>
 
 </Project>
