+++ Identity/SciMaterials.Identity.Api/appsettings.json
@@ -5,5 +5,18 @@
       "Microsoft.AspNetCore": "Warning"
     }
   },
-  "AllowedHosts": "*"
+  "AllowedHosts": "*",
+  "ConnectionStrings": {
+    "SqlServer.Debug": "Server=localhost;Database=SciMaterials.Identity;User Id=sa;Password=veryStrongPassword123;"
+  },
+  "DatabaseSettings": {
+    "Provider": "SqlServer.Debug",
+    "RemoveAtStart": "true",
+    "UseDataSeeder": "true"
+  },
+  "SecretTokenKey": "vRguqNCvKJHkhs578443sa46@#sakjjRxDGMFguqNCMz8w2DGMQ==",
+  "AdminSettings": {
+    "login": "sa@mail.ru",
+    "password": "test12345"
+  }
 }
